refactor: split CLI and parser helpers
