+++ src/bin.ts
@@ -44,8 +44,8 @@ type RunArgs = {
 };
 
 /* eslint-disable security/detect-non-literal-fs-filename */
-export async function main(argv: string[]) {
-	const args = mri(argv.slice(2), {
+function parseCliArgs(argv: string[]) {
+	return mri(argv.slice(2), {
 		alias: {
 			c: 'cache',
 			f: 'force',
@@ -54,87 +54,105 @@ export async function main(argv: string[]) {
 		},
 		boolean: ['force', 'cache', 'verbose'],
 	}) as CliArgs;
+}
+
+function displayHelp() {
+	const help = fs
+		.readFileSync(path.join(__dirname, '..', 'assets', 'help.md'), 'utf8')
+		.replaceAll(/^(\s*)#+ (.+)/gm, (match, indent, title) => indent + chalk.bold(title))
+		.replaceAll(/_([^_]+)_/g, (match, value) => chalk.underline(value))
+		.replaceAll(/`([^`]+)`/g, (match, value) => chalk.cyan(value));
+
+	process.stdout.write(`\n${help}\n`);
+}
+
+function getInteractiveChoices(): Choice[] {
+	const accessLookup = new Map<string, number>();
+
+	glob('**/access.json', { cwd: base }).forEach((file) => {
+		const normalizedFile = file.replaceAll('\\', '/');
+		const [host, user, repo] = normalizedFile.split('/');
+		const json = fs.readFileSync(`${base}/${file}`, 'utf8');
+		const logs = JSON.parse(json) as Record<string, string | number>;
+
+		Object.entries(logs).forEach(([ref, timestamp]) => {
+			const id = `${host}:${user}/${repo}#${ref}`;
+			accessLookup.set(id, new Date(String(timestamp)).getTime());
+		});
+	});
+
+	const getChoices = (file: string): Choice[] => {
+		const normalizedFile = file.replaceAll('\\', '/');
+		const [host, user, repo] = normalizedFile.split('/');
+		const entries = Object.entries(tryRequire(`${base}/${file}`) || {});
+
+		return entries.map(([ref, hash]) => ({
+			message: `${host}:${user}/${repo}#${ref}`,
+			name: String(hash),
+			value: `${host}:${user}/${repo}#${ref}`,
+		}));
+	};
+
+	return glob('**/map.json', { cwd: base })
+		.flatMap(getChoices)
+		.sort((a, b) => (accessLookup.get(b.value) || 0) - (accessLookup.get(a.value) || 0));
+}
+
+async function promptForSource(): Promise<PromptResult> {
+	const sourcePrompt = {
+		choices: getInteractiveChoices(),
+		message: 'Repo to clone?',
+		name: 'src',
+		suggest: (input: string, promptChoices: Choice[]) =>
+			promptChoices.filter(({ value }) => fuzzysearch(input, value)),
+		type: 'autocomplete',
+	} as any;
+
+	return enquirer.prompt<PromptResult>([
+		sourcePrompt,
+		{
+			initial: '.',
+			message: 'Destination directory?',
+			name: 'dest',
+			type: 'input',
+		},
+		{
+			message: 'Use cached version?',
+			name: 'cache',
+			type: 'toggle',
+		},
+	] as any);
+}
+
+async function confirmOverwrite(): Promise<boolean> {
+	const { force } = await enquirer.prompt<ForceResult>([
+		{
+			message: 'Overwrite existing files?',
+			name: 'force',
+			type: 'toggle',
+		},
+	] as any);
+
+	return force;
+}
+
+export async function main(argv: string[]) {
+	const args = parseCliArgs(argv);
 
 	const [src, dest = '.'] = args._;
 
 	if (args.help) {
-		const help = fs
-			.readFileSync(path.join(__dirname, '..', 'assets', 'help.md'), 'utf8')
-			.replaceAll(/^(\s*)#+ (.+)/gm, (match, indent, title) => indent + chalk.bold(title))
-			.replaceAll(/_([^_]+)_/g, (match, value) => chalk.underline(value))
-			.replaceAll(/`([^`]+)`/g, (match, value) => chalk.cyan(value));
-
-		process.stdout.write(`\n${help}\n`);
+		displayHelp();
 		return;
 	}
 
 	if (!src) {
-		const accessLookup = new Map<string, number>();
-
-		glob('**/access.json', { cwd: base }).forEach((file) => {
-			const normalizedFile = file.replaceAll('\\', '/');
-			const [host, user, repo] = normalizedFile.split('/');
-			const json = fs.readFileSync(`${base}/${file}`, 'utf8');
-			const logs = JSON.parse(json) as Record<string, string | number>;
-
-			Object.entries(logs).forEach(([ref, timestamp]) => {
-				const id = `${host}:${user}/${repo}#${ref}`;
-				accessLookup.set(id, new Date(String(timestamp)).getTime());
-			});
-		});
-
-		const getChoices = (file: string): Choice[] => {
-			const normalizedFile = file.replaceAll('\\', '/');
-			const [host, user, repo] = normalizedFile.split('/');
-			const entries = Object.entries(tryRequire(`${base}/${file}`) || {});
-
-			return entries.map(([ref, hash]) => ({
-				message: `${host}:${user}/${repo}#${ref}`,
-				name: String(hash),
-				value: `${host}:${user}/${repo}#${ref}`,
-			}));
-		};
-
-		const choices = glob('**/map.json', { cwd: base })
-			.flatMap(getChoices)
-			.sort((a, b) => (accessLookup.get(b.value) || 0) - (accessLookup.get(a.value) || 0));
-
-		const sourcePrompt = {
-			choices,
-			message: 'Repo to clone?',
-			name: 'src',
-			suggest: (input: string, promptChoices: Choice[]) =>
-				promptChoices.filter(({ value }) => fuzzysearch(input, value)),
-			type: 'autocomplete',
-		} as any;
-
-		const options = await enquirer.prompt<PromptResult>([
-			sourcePrompt,
-			{
-				initial: '.',
-				message: 'Destination directory?',
-				name: 'dest',
-				type: 'input',
-			},
-			{
-				message: 'Use cached version?',
-				name: 'cache',
-				type: 'toggle',
-			},
-		] as any);
+		const options = await promptForSource();
 
 		const empty = !fs.existsSync(options.dest) || fs.readdirSync(options.dest).length === 0;
 
 		if (!empty) {
-			const { force } = await enquirer.prompt<ForceResult>([
-				{
-					message: 'Overwrite existing files?',
-					name: 'force',
-					type: 'toggle',
-				},
-			] as any);
-
-			if (!force) {
+			if (!(await confirmOverwrite())) {
 				console.error(chalk.magenta('! Directory not empty — aborting'));
 				return;
 			}
